Adds argonaut package for Argo CD management
Adds the `argonaut` package, providing a terminal UI for managing Argo CD applications.

This allows users to interact with Argo CD in a more efficient and keyboard-driven manner.

Updates the `flake.nix` to include a custom packages overlay
and makes `argonaut` available for Darwin systems.
The `home.nix` file is also updated to include `argonaut` in the list of installed packages.